Refactor: move to cases.test.js
